# PERF: Introduce particle data structure for PSO