… libautoscoper/src/Tracker.cpp