* For any given particle p, this update removes any recomputation of the NCC fitness value.
* Replaces arrays with vectors
* This drops the per frame opt time from ~9 to ~18 seconds to around ~3 to ~ 6 seconds
* Based off of #221 
    * It would appear that most of the time gained was from these API changes, however, there still is ~.5 sec per frame left on the table from adding the multithreading.