@NicerNewerCar 2023-10-26
Tested with CUDA and that performs optimizations in ~2 seconds. Will be interesting to see how the difference between OpenCL and CUDA progresses as we integrate this PR, the multithreading PR, and the PR to remove the repeated rendering of the radiograph.