@amymmorton 2023-10-26
wait - I'm confused - 

"For any given particle p, this update removes any recomputation of the NCC fitness value."

That's an integral part of pso we don't want to remove
I need to step through your implementation to make sure

At each iteration:

- For all particles- we want the ncc , 
- For all particles- we compare if this iteration's ncc is the best That specific particle has known
- The swarm best particle
- update the particle position and vector from above