@jcfr 2023-10-31
> Previously: We would have computed the NCC value for p 3-4 times
> Now: We are only computing the NCC for a given p once

In the previous implementation, there was a for loop iterating over all particles:

https://github.com/BrownBiomechanics/Autoscoper/blob/73eb359dcd70878c4b4c28471fd2684c08bd3014/libautoscoper/src/PSO.cpp#L83-L89

where we computing the fitness function 5 times:

* The first two calls to `host_fitness_function` where checking if the current particle N was better than the "best" of the current iteration.

  https://github.com/BrownBiomechanics/Autoscoper/blob/73eb359dcd70878c4b4c28471fd2684c08bd3014/libautoscoper/src/PSO.cpp#L91C11-L96


* The second two calls to `host_fitness_function` where checking if the current particle N was better for "global best"

  https://github.com/BrownBiomechanics/Autoscoper/blob/73eb359dcd70878c4b4c28471fd2684c08bd3014/libautoscoper/src/PSO.cpp#L98-L104

* And the last call to `host_fitness_function` was used to get the value for the updated global best:
 
  https://github.com/BrownBiomechanics/Autoscoper/blob/73eb359dcd70878c4b4c28471fd2684c08bd3014/libautoscoper/src/PSO.cpp#L109


As outlined by @NicerNewerCar, in the proposed update, we have a collection of `Particle` object in which we keep track of already computed NCC value.

This means we evaluate each particle once at every iteration instead of between 2 and 5 times (worst case).