@NicerNewerCar 2023-11-06
## Analysis of the new data structure

### Summary
The new data structure introduces a significant speed-up in optimization time while retaining similar accuracy to the previous method

### Experimental Setup

* WN00113
* Tracked 100 frames of the mc3 volume
* No seeding

### Optimization Time Results

```bash
Old PSO:
Average:  19.022589359283447
Median:  18.46156370639801
St.Dev.: 2.902542623887008

New PSO:
Average:  5.78880467414856
Median:  5.697943687438965
St.Dev.: 0.6337049489363932
```
### Accuracy 

![old_pso](https://github.com/BrownBiomechanics/Autoscoper/assets/30351234/f03ac57b-02d0-47d8-bbe1-a0fb24d9788b)
![new_pso](https://github.com/BrownBiomechanics/Autoscoper/assets/30351234/afce69d2-822e-49f2-806b-e27e8c052e0b)