@jcfr 2023-11-06
@amymmorton Once you are done reviewing, we will move forward with the integration.

For reference, the commit message suggested in the last comment[^1] will be used when integration using `Squash & Merge`.

[^1]: https://github.com/BrownBiomechanics/Autoscoper/pull/223#issuecomment-1796380246