@jcfr 2023-11-06
> I notice that this is CUDA back-end - was that intentional/ required?

The change implemented in this pull request aims to be back-end agnostic. This means that expects improvement in both cases.

That said, we should indeed update the commit message to be more specific when describing the performance gain. I am referring to this line:

> This optimization has led to a remarkable reduction in per-frame execution time
on a typical workstation, dropping it from the range of 9 to 18 seconds to a
much-improved 3 to 6 seconds.

@NicerNewerCar In https://github.com/BrownBiomechanics/Autoscoper/pull/223#issuecomment-1781551387, you mention that the changes were tested against the CUDA back-end.

**Question:** Is that also true in the context of the analysis documented in https://github.com/BrownBiomechanics/Autoscoper/pull/223#issuecomment-1795927981 ? Was it performed with the CUDA back-end .. and if not would be nice to also have an understanding on how the PSO refactoring impacts both.