@amymmorton 2023-11-06
built and tested as was CUDA: tracking 10 seeded frames rad_uln WN00108
 about 2.5 sec per frame

built openCL - same 10 frames (same params)  approx 12.5 sec per frame

On Mon, Nov 6, 2023 at 4:27 PM Jean-Christophe Fillion-Robin <
***@***.***> wrote:

> I notice that this is CUDA back-end - was that intentional/ required?
>
> The change implemented in this pull request aims to be back-end agnostic.
> This means that expects improvement in both cases.
>
> That said, we should indeed update the commit message to be more specific
> when describing the performance gain. I am referring to this line:
>
> This optimization has led to a remarkable reduction in per-frame execution
> time
> on a typical workstation, dropping it from the range of 9 to 18 seconds to
> a
> much-improved 3 to 6 seconds.
>
> @NicerNewerCar <https://github.com/NicerNewerCar> In #223 (comment)
> <https://github.com/BrownBiomechanics/Autoscoper/pull/223#issuecomment-1781551387>,
> you mention that the changes were tested against the CUDA back-end.
>
> *Question:* Is that also true in the context of the analysis documented
> in #223 (comment)
> <https://github.com/BrownBiomechanics/Autoscoper/pull/223#issuecomment-1795927981>
> ? Was it performed with the CUDA back-end .. and if not would be nice to
> also have an understanding on how the PSO refactoring impacts both.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/BrownBiomechanics/Autoscoper/pull/223#issuecomment-1796518327>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/APUUPFSWHT2FKZLHBXDZAGLYDFI2VAVCNFSM6AAAAAA6RQMJZGVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTOOJWGUYTQMZSG4>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>


-- 
-----------
Amy Morton, MSc

Sr. Research Engineer, Bioengineering Lab, Department of Orthopaedic
Research
The Warren Alpert Medical School of Brown University and Rhode Island
Hospital
1 Hoppin Street, CORO West, Ste. 404, Providence, RI 02903