@jcfr 2023-11-07
@amymmorton Thanks for testing, we will now move forward with the integration :rocket: